migrated helper and connector test to pytest
